test(RaceCars): 우승자 테스트
