+++ src/test/java/domain/RaceCarsTest.java
@@ -6,6 +6,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -41,15 +42,25 @@ class RaceCarsTest {
     }
 
     @Test
-    @DisplayName("Car 리스트를 반환한다.")
-    void getCars() {
-        List<Car> cars = List.of(
-                new Car("hi"),
-                new Car("hello")
-        );
-
+    @DisplayName("가장 멀리 이동한 자동차가 우승한다")
+    void getWinners() {
+        //given
+        Car one = new Car("one");
+        Car two = new Car("two");
+        Car three = new Car("three");
+        List<Car> cars = new ArrayList<>(List.of(one, two, three));
         RaceCars raceCars = new RaceCars(cars);
 
-        assertThat(raceCars.getCars()).isEqualTo(cars);
+        one.move(4);
+        two.move(0);
+        three.move(8);
+
+        RacingGame racingGame = new RacingGame(raceCars, new Count(1));
+
+        //when
+        List<Car> winners = racingGame.judgeWinners();
+
+        //then
+        assertThat(winners).contains(one, three);
     }
 }
